KL1/src/backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.7
pyrad==2.1
bcrypt==4.0.1
Jinja2==3.1.2
requests==2.31.0
cryptography==41.0.7
xmltodict==0.13.0
pyOpenSSL==23.2.0
netaddr==0.8.0
psutil==5.9.5
gunicorn==21.2.0
python-dateutil==2.8.2
pyyaml==6.0.1
sqlalchemy==2.0.23
alembic==1.12.1